cloud/atot/requirements.txt
openai-whisper
ffmpeg
numpy>=1.21.0,<2.0.0